Temp commit at 25 April
Going to change from FIFO to Ram. Do to fifo doesn't support the length
of 160 and fifo doesn't support trigger. Have found the tutorial.

To the final, I have finished to draw by using bit shift and add a
consistent of -4. There is the sine wave at ~500khz(Matlab script created
to describe the sinewave). Trigger is not suppport. The plot is breaked at
some point due to the poor fifo and state machine.

The first thing tomorrow is to implement the ram so I can then see a
continuous plot. Then the trigger to see a static image, finial the
compressor to be able to manipulate the image.